video_creator_project/requirements.txt
google-generativeai==0.3.1
deep-translator==1.11.4
gTTS==2.3.2
moviepy==1.0.3
Pillow==10.0.0
pydub==0.25.1
torch==2.0.1
torchvision==0.15.2
diffusers==0.19.3
transformers==4.31.0
accelerate==0.21.0